requirements.txt
numpy==1.19.5
tensorboard
tqdm
ConfigArgParse
matplotlib
plotly
shapely
scipy
setuptools
scikit-learn
pillow
pyyaml
pandas
joblib
sentence_transformers
wandb
trimesh
mesh_to_sdf
seaborn
tabulate
einops
torchdiffeq
torch-summary
pygame
gitpython
scikit-video
functorch
opencv-python
moviepy
hydra
urdfpy
gdown
